backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
PyJWT==2.9.0
pymongo==4.11.2
sqlparse==0.5.3
tzdata==2025.1
django-environ~=0.12.0